reorganize the typia-related code
