+++ .gitignore
@@ -14,6 +14,7 @@ dist/*
 server/dataset/*.js
 server/genome/*.js
 server/shared/doc.js
+server/shared/checkers/transformed/*
 server/shared/types/docs.json
 server/shared/types/*.js
 !server/shared/types/typedoc.js
