+++ public/testrun.html
@@ -48,7 +48,7 @@
   <span id='pp-test-runner-banner' class='code-snippet'>$ npm run dev</span> 
   must be active to (re)compile 
   <span class='code-snippet'>*.spec.js</span> files.
-  <span>For faster repeated reload of unchanged test(s), add a '&reuse=1' parameter to the URL.</span>
+  <span>For faster repeated reload of unchanged test code, add a <span class='code-snippet'>reuse=1</span> parameter to the URL.</span>
 </div>
 <div style='position: absolute; right: 20px; margin: 20px; font-family: Menlo, Courier;'>
   <button><a href='testrun.html?name=*.unit'>All unit tests</a></button>
