+++ server/package.json
@@ -20,8 +20,8 @@
     "prepack": "tsc --esModuleInterop genome/*.ts dataset/*.ts && webpack --env NODE_ENV=production",
     "response": "nodemon modules/test/test.server.js --watch src",
     "getconf": "../build/getConfigProp.js",
-    "doc": "typedoc",
-    "gentesters": "typia generate --input types/testers --output test/testers --project types/testers/tsconfig.json",
+    "doc": "typedoc --json ../public/docs/server.json",
+    "gentesters": "typia generate --input shared/checkers/raw --output shared/checkers/transformed --project ./shared/checkers/tsconfig.json",
     "doctest": "npm run gentesters && npm run test:unit"
   },
   "author": "",
