+++ server/shared/checkers/raw/index.ts
@@ -1,5 +1,5 @@
 import { createIs, createValidate } from 'typia'
-import { HealthCheckResponse } from '../../src/health'
+import { HealthCheckResponse } from '../../types/healthcheck'
 
 export const isHealthCheckResponse = createIs<HealthCheckResponse>()
 export const validHealthCheckResponse = createValidate<HealthCheckResponse>()
